requirements_conda.txt
# Create an environment using: conda create --name <env> --file <this file> -c pytorch
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
torchvision=0.12.0=py37_cu113
torchaudio=0.11.0=py37_cu113
h5py=3.6.0
python=3.7.13
matplotlib=3.5.1
numpy=1.21.2
scikit-learn=0.24.2
scipy=1.7.3
